+++ packages/models/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Remove box.updateImage
 - Implement deletion of box images
 - Allow updating of box model value
+- Add 'includeSecrets' to user model toJSON
 
 ## v0.0.9
 - Fix measurements with 'content-type: json' not parsed with JSON.parse.